Move Mojang API to separate file
